+++ single-wpbb_company.php
@@ -1,0 +1,1 @@
+<?php defined( 'ABSPATH' ) || exit; get_header(); echo wpbb_jobs_single_company_view(); /* phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped */ get_footer();